+++ lib/factual/write/flag.rb
@@ -1,9 +1,7 @@
 class Factual
   module Write
     class Flag < Base
-      PROBLEMS = [:closed, :duplicate, :nonexistent, :inaccurate, :inappropriate, :spam, :relocated, :other]
-      VALID_KEYS = [:table, :factual_id, :problem, :user, :comment, :debug, :reference, :fields, :preferred]
-      DEPRECATED_KEYS = [:data]
+      VALID_KEYS = [:table, :factual_id, :data, :problem, :user, :comment, :debug, :reference, :fields, :preferred]
 
       def initialize(api, params)
         validate_params(params)
@@ -24,12 +22,7 @@ class Factual
 
       def validate_params(params)
         params.keys.each do |key|
-          raise "Deprecated flag option: #{key}" if DEPRECATED_KEYS.include?(key)
           raise "Invalid flag option: #{key}" unless VALID_KEYS.include?(key)
-        end
-
-        unless PROBLEMS.include?(params[:problem])
-          raise "Flag problem should be one of the following: #{PROBLEMS.join(", ")}"
         end
       end
     end